+++ sdk/dependencymap/azure-mgmt-dependencymap/generated_samples/maps_get_connections_with_connected_machine_for_focused_machine.py
@@ -8,7 +8,7 @@
 
 from azure.identity import DefaultAzureCredential
 
-from azure.mgmt.dependencymap import DependencyMapMgmtClient
+from azure.mgmt.dependencymap import DependencyMapClient
 
 """
 # PREREQUISITES
@@ -25,7 +25,7 @@ from azure.mgmt.dependencymap import DependencyMapMgmtClient
 
 
 def main():
-    client = DependencyMapMgmtClient(
+    client = DependencyMapClient(
         credential=DefaultAzureCredential(),
         subscription_id="SUBSCRIPTION_ID",
     )
@@ -47,6 +47,6 @@ def main():
     ).result()
 
 
-# x-ms-original-file: 2025-01-31-preview/Maps_GetConnectionsWithConnectedMachineForFocusedMachine.json
+# x-ms-original-file: 2025-07-01-preview/Maps_GetConnectionsWithConnectedMachineForFocusedMachine.json
 if __name__ == "__main__":
     main()