+++ sdk/dependencymap/azure-mgmt-dependencymap/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Release History
 
+## 1.0.0b2 (2025-09-23)
+
+### Features Added
+
+  - Model `ExportDependenciesRequest` added property `appliance_name_list`
+  - Added model `DependencyProcessFilter`
+  - Added model `ExportDependenciesAdditionalInfo`
+  - Added model `ExportDependenciesOperationResult`
+  - Added model `ExportDependenciesResultProperties`
+  - Added enum `ExportDependenciesStatusCode`
+  - Added model `GetDependencyViewForAllMachinesOperationResult`
+  - Added model `GetDependencyViewForAllMachinesRequest`
+  - Added model `GetDependencyViewForAllMachinesResultProperties`
+  - Model `MapsOperations` added method `begin_get_dependency_view_for_all_machines`
+
+### Breaking Changes
+
+  - Deleted or renamed client `DependencyMapMgmtClient`
+
 ## 1.0.0b1 (2025-04-20)
 
 ### Other Changes